backend/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "event-photo-gallery-backend"
version = "0.1.0"
description = "Backend API for Personal Event Photo Gallery with Smart Face Recognition"
authors = [{name = "Event Gallery Team"}]
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.12.1",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.6",
    "face-recognition>=1.3.0",
    "pillow>=10.1.0",
    "redis>=5.0.1",
    "boto3>=1.34.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "httpx>=0.25.2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"